# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: EdgeAI Control Plane
  version: 0.1.0
  description: Profile → Device/Node → Workflow DAG → Run/Task/Attempt와 검증된 결과 조회 API입니다.
    실제 Kubernetes 실행은 서버의 runtime 설정이 활성화된 환경에서 동작합니다. 쓰기 요청은 실제 DB에 반영되고 Swagger의 CSRF 토큰/쿠키는 자동 연결됩니다.
    복원 DB의 조회 전용 점검 모드에서는 실행 기능이 비활성화되고 CSRF가 유효한 쓰기 요청도 403으로 거절됩니다.
    일반 관리 API의 변경 요청은 먼저 감사 접수 기록을 저장하며 X-EdgeAI-Audit-Id를 반환합니다. 접수 저장 실패는503/AUDIT_STORE_UNAVAILABLE이며 변경을 실행하지 않습니다.
    감사의 HTTP 결과는 비동기 작업 성공을 뜻하지 않으며, 결과가 없는 접수 기록은 OUTCOME_UNKNOWN입니다. 복구 점검 모드에서는 감사 쓰기도 비활성화됩니다.
servers:
- url: /
security: []
paths:
...
  /api/v1/devices:
    post:
      tags:
      - Device
      operationId: registerDevice
      summary: 물리 장치 등록
      description: DEVICE Profile의 발행 UUID를 참조해 장치를 등록합니다. 같은 key와 동일 생성 입력은 200, 새로운 장치는 201, 다른 생성 입력은 409입니다. LIVE/REPLAY/SYNTHETIC으로
        데이터 출처를 구분합니다. 등록만으로 실제 연결을 ONLINE으로 표시하지 않습니다. 요청 최대 16 KiB.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '201':
          description: 새 객체 생성. 기존 객체의 ID는 바뀌지 않습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
    get:
      tags:
      - Device
      operationId: listDevices
      summary: 장치 목록과 연결 상태 조회
      description: key 문자열 오름차순으로 조회합니다. 해제된 장치도 이력 조회를 위해 포함합니다. 활성 세션의 최근 보고가 60초보다 오래되면 STALE이며, 보고 전에는 UNKNOWN입니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePage'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: 최대 항목 수
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          maximum: 1000000
          default: 0
        description: nextOffset을 전달해 다음 페이지를 조회합니다.
  /api/v1/devices/{deviceId}:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Device
      operationId: getDevice
      summary: 장치 상세·연결 이력·최근 관측 조회
      description: 장치와 최근 연결/세션/관측 각각 최대 20개를 반환합니다. 시간은 UTC이며 신선도는 서버 수신 시각으로 판단합니다. raw 스트림은 저장하지 않습니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Device
      operationId: updateDevice
      summary: 장치 표시 이름 수정
      description: 상세 조회에서 받은 revision과 새 displayName을 전송합니다. 동시에 수정되어 revision이 바뀌면 409입니다. Profile 참조·key·sourceMode는 변경하지
        않습니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicePatch'
    delete:
      tags:
      - Device
      operationId: releaseDevice
      summary: 장치 해제 및 세션 종료
      description: 장치를 RELEASED로 표시하고 활성 attachment/session을 닫습니다. 등록된 VD 또는 종료 대기 중인 이전 VD 실행이 원본으로 사용하면 DEVICE_IN_USE(409)입니다. 원본 연결 해제와 이전 실행의 종료 확인이 모두 필요합니다. 과거 이력을 보존하며 반복 해제는 같은 상태를 반환합니다. 해제 후 연결·세션·관측·수정 요청은 409입니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
  /api/v1/devices/{deviceId}/registration:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema: {type: string, format: uuid}
    delete:
      operationId: deleteDeviceRegistration
      tags: [Device]
      summary: 사용하지 않는 장치 등록 삭제
      description: 활성 연결·세션이나 VD·실행 이력 참조가 있으면 DEVICE_IN_USE(409)로 거절합니다. 미사용 장치와 그 장치의 닫힌 연결·세션·관측 이력만 삭제하며 Kubernetes 노드는 변경하지 않습니다.
      security:
      - csrfToken: []
      responses:
        '204':
          description: 장치 등록 삭제 완료
        '400':
          $ref: '#/components/responses/Invalid'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: DEVICE_IN_USE — 활성 연결·세션 또는 다른 자원의 참조가 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/devices/{deviceId}/attachments/{nodeId}:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    - in: path
      name: nodeId
      required: true
      schema:
        type: string
        format: uuid
    put:
      tags:
      - Device
      operationId: attachDevice
      summary: 장치를 관측된 실행 노드에 연결
      description: 최근 60초 안에 실제 관측된 Ready Node와 port를 연결합니다. 기존 활성 연결은 닫고 새 이력을 만들며 동일 node/port 재전송은 이력을 추가하지 않습니다. 장치당 활성
        연결은 하나입니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAttachment'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAttachmentInput'
  /api/v1/devices/{deviceId}/sessions:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    post:
      tags:
      - Device
      operationId: openDeviceSession
      summary: 재접속·재부팅 세션 시작
      description: bootId는 재부팅/재접속마다 새 UUID입니다. 현재 bootId 재전송은 200, 새 bootId는 기존 세션을 닫고 epoch를 올려 201입니다. 닫힌 bootId는 재사용할
        수 없습니다. 이전 세션의 관측은 차단됩니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSession'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '201':
          description: 새 객체 생성. 기존 객체의 ID는 바뀌지 않습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSession'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceSessionInput'
  /api/v1/devices/{deviceId}/observations:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    post:
      tags:
      - Device
      operationId: reportDeviceObservation
      summary: 장치 상태·작은 관측 데이터 보고
      description: 현재 sessionId, 증가하는 sequence, observedAt, ONLINE/OFFLINE, attributes를 보냅니다. 동일 sequence/내용 재전송은 200, 내용
        변경·역순·이전 세션은 409입니다. observedAt은 서버보다 최대 30초 미래·24시간 과거를 허용합니다. 요청 최대 16 KiB이며 raw stream/대형 파일은 대상이 아닙니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceObservation'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '201':
          description: 새 객체 생성. 기존 객체의 ID는 바뀌지 않습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceObservation'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceObservationInput'
...
components:
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 쓰기 요청은 세션 쿠키와 함께 이 헤더가 필요합니다. Swagger는 자동 발급·전달하므로 이 입력칸을 비워 두세요. 직접 호출할 때는 getCsrfToken을 먼저 사용합니다.
...
  responses:
...
    Invalid:
      description: 종류·키·버전·페이지·JSON 형식이 잘못되었습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_PROFILE
            message: 종류, 키, 버전과 JSON 규격을 확인하세요. 규격은 비어 있지 않은 객체여야 합니다.
...
    NotFound:
      description: 해당 발행 버전이 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_NOT_FOUND
            message: 해당 Profile 버전을 찾을 수 없습니다.
    Unavailable:
      description: DB 연결 등 저장소 문제로 요청을 처리할 수 없습니다. 복구 후 재시도하세요.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_STORE_UNAVAILABLE
            message: Profile 저장소에 연결할 수 없습니다. 잠시 후 다시 시도하세요.
  schemas:
...
    ProfileKey:
      type: string
      maxLength: 100
      pattern: ^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$
      description: 소문자 영문으로 시작하는 프로필 키. 영문 소문자·숫자와 구분자 점/밑줄/하이픈을 사용하며 최대 100자입니다.
      example: temperature-sensor
...
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_PROFILE
          - PROFILE_CONFLICT
          - PROFILE_NOT_FOUND
          - PROFILE_STORE_UNAVAILABLE
          - PAYLOAD_TOO_LARGE
          - INVALID_DEVICE
          - DEVICE_CONFLICT
          - DEVICE_NOT_FOUND
          - NODE_NOT_FOUND
          - NODE_NOT_READY
          - DEVICE_RELEASED
          - STALE_SESSION
          - OBSERVATION_CONFLICT
          - DEVICE_STORE_UNAVAILABLE
          - INVALID_WORKFLOW
          - WORKFLOW_NOT_FOUND
          - WORKFLOW_CONFLICT
          - RUN_NOT_FOUND
          - TASK_NOT_FOUND
          - IDEMPOTENCY_CONFLICT
          - CANNOT_CANCEL
          - STREAM_NOT_IMPLEMENTED
          - STREAM_EXECUTION_POLICY_UNSUPPORTED
          - STREAM_RECOVERY_UNSUPPORTED
          - STREAM_SOURCE_INACTIVE
          - WORKFLOW_STORE_UNAVAILABLE
          - INVALID_TASK_ID
          - RESULT_STORE_UNAVAILABLE
          - RUNTIME_DISABLED
          - REMOTE_DISABLED
          - REMOTE_PROVIDER_NOT_FOUND
          - REMOTE_CONFIGURATION_CHANGED
          - REMOTE_TELEMETRY_UNSUPPORTED
          - VD_EXECUTION_DISABLED
          - VD_NOT_READY
          - VD_SERVICE_MISMATCH
          - VD_AUTOMATIC_OFFLOAD_UNSUPPORTED
          - OFFLOAD_SOURCE_CHANGED
          - OFFLOAD_LIMIT
          - OFFLOAD_RECOVERY_UNSUPPORTED
          - OFFLOAD_TARGET_INVALID
          - OPERATION_NOT_FOUND
          - INVALID_VIRTUAL_DEVICE
          - VD_NOT_FOUND
          - VD_CONFLICT
          - VD_RELEASED
          - VD_SOURCE_INCOMPATIBLE
          - VD_STORE_UNAVAILABLE
          - DEVICE_IN_USE
        message:
          type: string
...
    DeviceCreate:
      type: object
      additionalProperties: false
      required:
      - key
      - displayName
      - profileVersionId
      - sourceMode
      properties:
        key:
          $ref: '#/components/schemas/ProfileKey'
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        profileVersionId:
          type: string
          format: uuid
        sourceMode:
          type: string
          enum:
          - LIVE
          - REPLAY
          - SYNTHETIC
    DevicePatch:
      type: object
      additionalProperties: false
      required:
      - revision
      - displayName
      properties:
        revision:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        displayName:
          type: string
          minLength: 1
          maxLength: 128
    DeviceAttachmentInput:
      type: object
      additionalProperties: false
      required:
      - port
      properties:
        port:
          type: string
          minLength: 1
          maxLength: 128
    DeviceSessionInput:
      type: object
      additionalProperties: false
      required:
      - bootId
      properties:
        bootId:
          type: string
          format: uuid
    DeviceObservationInput:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - sequence
      - observedAt
      - status
      - attributes
      properties:
        sessionId:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        observedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - ONLINE
          - OFFLINE
        attributes:
          type: object
          additionalProperties: true
    Device:
      type: object
      additionalProperties: false
      required:
      - id
      - key
      - displayName
      - profileVersionId
      - sourceMode
      - state
      - revision
      - sessionEpoch
      - connectionStatus
      - lastSeenAt
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        key:
          $ref: '#/components/schemas/ProfileKey'
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        profileVersionId:
          type: string
          format: uuid
        sourceMode:
          type: string
          enum:
          - LIVE
          - REPLAY
          - SYNTHETIC
        state:
          type: string
          enum:
          - ACTIVE
          - RELEASED
        revision:
          type: integer
          minimum: 0
        sessionEpoch:
          type: integer
          minimum: 0
        connectionStatus:
          type: string
          enum:
          - UNKNOWN
          - ONLINE
          - OFFLINE
          - STALE
          - RELEASED
        lastSeenAt:
          type:
          - string
          - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceAttachment:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - nodeId
      - port
      - attachedAt
      - detachedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        nodeId:
          type: string
          format: uuid
        port:
          type: string
        attachedAt:
          type: string
          format: date-time
        detachedAt:
          type:
          - string
          - 'null'
          format: date-time
    DeviceSession:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - bootId
      - epoch
      - lastSequence
      - openedAt
      - closedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        bootId:
          type: string
          format: uuid
        epoch:
          type: integer
          minimum: 1
        lastSequence:
          type: integer
          minimum: -1
        openedAt:
          type: string
          format: date-time
        closedAt:
          type:
          - string
          - 'null'
          format: date-time
    DeviceObservation:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - sessionId
      - sequence
      - observedAt
      - status
      - attributes
      - receivedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        observedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - ONLINE
          - OFFLINE
        attributes:
          type: object
          additionalProperties: true
        receivedAt:
          type: string
          format: date-time
    DeviceDetail:
      type: object
      additionalProperties: false
      required:
      - device
      - attachments
      - sessions
      - observations
      properties:
        device:
          $ref: '#/components/schemas/Device'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/DeviceAttachment'
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/DeviceSession'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/DeviceObservation'
...
    DevicePage:
      type: object
      additionalProperties: false
      required:
      - items
      - nextOffset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        nextOffset:
          type:
          - integer
          - 'null'
...